Refactor web progress source
