+++ apps/web/src/appData/progressTimeContext.ts
@@ -1,11 +1,12 @@
 import { useSyncExternalStore } from "react";
+import {
+  buildProgressDateContext,
+  type ProgressDateContext,
+} from "../progress/progressDates";
 
 const progressTimeContextPollIntervalMs = 60_000;
 
-export type ProgressTimeContext = Readonly<{
-  timeZone: string;
-  today: string;
-}>;
+export type ProgressTimeContext = ProgressDateContext;
 
 type ProgressTimeContextListener = () => void;
 
@@ -13,51 +14,8 @@ let progressTimeContextSnapshot: ProgressTimeContext | null = null;
 
 const progressTimeContextListeners = new Set<ProgressTimeContextListener>();
 
-function getRequiredDatePart(
-  parts: ReadonlyArray<Intl.DateTimeFormatPart>,
-  partType: "year" | "month" | "day",
-): string {
-  const partValue = parts.find((part) => part.type === partType)?.value;
-
-  if (partValue === undefined || partValue === "") {
-    throw new Error(`Browser timezone date is missing ${partType}`);
-  }
-
-  return partValue;
-}
-
-function getBrowserTimeZone(): string {
-  const timeZone = Intl.DateTimeFormat().resolvedOptions().timeZone;
-
-  if (typeof timeZone !== "string" || timeZone.trim() === "") {
-    throw new Error("Browser timezone is unavailable");
-  }
-
-  return timeZone;
-}
-
-function formatDateAsLocalDate(date: Date, timeZone: string): string {
-  const formatter = new Intl.DateTimeFormat("en-CA", {
-    timeZone,
-    year: "numeric",
-    month: "2-digit",
-    day: "2-digit",
-  });
-  const parts = formatter.formatToParts(date);
-  const year = getRequiredDatePart(parts, "year");
-  const month = getRequiredDatePart(parts, "month");
-  const day = getRequiredDatePart(parts, "day");
-
-  return `${year}-${month}-${day}`;
-}
-
 export function buildProgressTimeContext(now: Date): ProgressTimeContext {
-  const timeZone = getBrowserTimeZone();
-
-  return {
-    timeZone,
-    today: formatDateAsLocalDate(now, timeZone),
-  };
+  return buildProgressDateContext(now);
 }
 
 function areProgressTimeContextsEqual(
